Week 2/Day 5/Hackathon/requirements.txt
psycopg2
psycopg2-binary
flask
requests
python-dotenv
jsonify